ART: Faster 712-varhandle-invocations
Reduce number of allocations when running 712-varhandle-invocations as
it timeouts under gcstress.

In the runtime, avoid allocating a MethodType when raising a
WrongMethodTypeException when dispatching an erroneous VarHandle
accessor.

In the test, limit the number of incorrect types tested in boxing test
portion of 712 which is particularly slow. And pre-allocate boxed
values and share across sub-tests.

The total time to run 712-varhandle-invocations is reduced by 45% on
host and 33% on angler.

Test: art/test/run-test --host --64 --gcstress 712
Bug: 73275005
Change-Id: If5b323a61291d490f51638d416c2529874282f1c
Signed-off-by: mydongistiny <jaysonedson@gmail.com>